feat: tests to work with dev CLI versions
