+++ acceptance-tests/src/test/scala/ch/renku/acceptancetests/workflows/CLIConfiguration.scala
@@ -19,6 +19,7 @@
 package ch.renku.acceptancetests.workflows
 
 import ch.renku.acceptancetests.model.CliVersion
+import ch.renku.acceptancetests.model.CliVersion.{NonReleasedVersion, ReleasedVersion}
 import ch.renku.acceptancetests.model.users.UserCredentials
 import ch.renku.acceptancetests.tooling.TestLogger.logger
 import ch.renku.acceptancetests.tooling.console._
@@ -32,8 +33,8 @@ trait CLIConfiguration {
   def `setup renku CLI`: Unit = {
     implicit val workFolder: Path = rootWorkDirectory
 
-    if (apiCliVersion != cliVersion)
-      logger.warn(s"Local Renku CLI is in version $cliVersion while Renku API is on $apiCliVersion")
+    if (apiCliVersion.value != cliVersion.value)
+      logger.warn(s"Local Renku CLI is is going to use v$cliVersion while Renku API is on v$apiCliVersion")
 
     CliVersion.get(console %%> c"renku --version") match {
       case Right(`cliVersion`) =>
@@ -41,8 +42,14 @@ trait CLIConfiguration {
       case _ =>
         Given(s"there's no Renku CLI $cliVersion installed")
         console %%> c"python3 -m pip uninstall --yes renku"
-        Then(s"the user installs Renku CLI $cliVersion")
-        console %> c"python3 -m pip install 'renku==$cliVersion' --user"
+        cliVersion match {
+          case version: ReleasedVersion =>
+            Then(s"the user installs released Renku CLI v$cliVersion")
+            console %> c"python3 -m pip install 'renku==$version'"
+          case version: NonReleasedVersion =>
+            Then(s"the user installs Renku CLI v$cliVersion from source")
+            console %> c"python3 -m pip install git+https://github.com/SwissDataScienceCenter/renku-python.git@${version.commitSha}"
+        }
     }
   }
